requirements.txt
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
numpy<2.0
scikit-learn>=1.0.0
Pillow>=8.0.0
requests>=2.25.0
